factor out unit test main
