+++ alica_test/src/test/test_main.cpp
@@ -1,0 +1,11 @@
+#include <gtest/gtest.h>
+#include <ros/ros.h>
+
+int main(int argc, char **argv)
+{
+	testing::InitGoogleTest(&argc, argv);
+	ros::init(argc, argv, "AlicaEngine");
+	bool ret = RUN_ALL_TESTS();
+	ros::shutdown();
+	return ret;
+}
